fixed the admin weeks and consolidated all time functions
